remove od versions sidebar
