+++ web_src/src/components/CanvasToolSidebar/EmptyToolTab.tsx
@@ -1,0 +1,3 @@
+export function EmptyToolTab() {
+  return <div className="min-h-0 flex-1" aria-hidden />;
+}
